+++ src/main/java/com/showmeyourcode/projects/algorithms/benchmark/BenchmarkDataGenerator.java
@@ -1,6 +1,7 @@
 package com.showmeyourcode.projects.algorithms.benchmark;
 
 import com.showmeyourcode.projects.algorithms.configuration.SortingAppConfiguration;
+import com.showmeyourcode.projects.algorithms.exception.BenchmarkDataNotFoundException;
 import com.showmeyourcode.projects.algorithms.generator.BaseDataGenerator;
 
 import java.io.BufferedInputStream;
@@ -9,6 +10,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
+import java.util.Objects;
 
 public class BenchmarkDataGenerator extends BaseDataGenerator {
 
@@ -16,19 +18,21 @@ public class BenchmarkDataGenerator extends BaseDataGenerator {
         super(appConfiguration);
     }
 
-    public int[] loadData(BenchmarkData benchmarkData) throws IOException {
+    public int[] loadData(BenchmarkData benchmarkData) throws BenchmarkDataNotFoundException, IOException {
         InputStream inputStream = getClass().getClassLoader().getResourceAsStream(benchmarkData.getPath());
 
-        BufferedInputStream bis = new BufferedInputStream(inputStream);
-        ByteArrayOutputStream buf = new ByteArrayOutputStream();
-        for (int result = bis.read(); result != -1; result = bis.read()) {
-            buf.write((byte) result);
+        if (Objects.isNull(inputStream)) {
+            throw new BenchmarkDataNotFoundException(benchmarkData);
+        } else {
+            BufferedInputStream bis = new BufferedInputStream(inputStream);
+            ByteArrayOutputStream buf = new ByteArrayOutputStream();
+            for (int result = bis.read(); result != -1; result = bis.read()) {
+                buf.write((byte) result);
+            }
+            return Arrays.stream(buf.toString(StandardCharsets.UTF_8).split(","))
+                    .map(String::trim)
+                    .map(Integer::parseInt)
+                    .mapToInt(i -> i).toArray();
         }
-        return Arrays.stream(buf.toString(StandardCharsets.UTF_8).split(","))
-                .map(String::trim)
-                .map(Integer::parseInt)
-                .mapToInt(i -> i).toArray();
     }
-
-
 }