+++ src/main/java/com/showmeyourcode/projects/algorithms/exception/BenchmarkDataNotFoundException.java
@@ -1,0 +1,9 @@
+package com.showmeyourcode.projects.algorithms.exception;
+
+import com.showmeyourcode.projects.algorithms.benchmark.BenchmarkData;
+
+public class BenchmarkDataNotFoundException extends Exception {
+    public BenchmarkDataNotFoundException(BenchmarkData benchmarkData) {
+        super("A Benchmark data does not exist! Benchmark: " + benchmarkData.getPath());
+    }
+}