+++ src/main/java/com/showmeyourcode/projects/algorithms/exception/CannotCreateReportResultsFileException.java
@@ -1,0 +1,7 @@
+package com.showmeyourcode.projects.algorithms.exception;
+
+public class CannotCreateReportResultsFileException extends Exception {
+    public CannotCreateReportResultsFileException(Throwable e) {
+        super(e);
+    }
+}